#26 removed vehicle tire - extending tire entity by vehicle tire fields + additional ones - problem with mapstruct
