+++ scaffolding-company-management-app/src/main/java/pl/com/chrzanowski/scma/domain/VehicleModel.java
@@ -28,6 +28,7 @@ public class VehicleModel {
 
     @ManyToOne
     @JoinColumn(name = "vehicle_brand")
+    @NotNull
     private VehicleBrand vehicleBrand;
 
     public VehicleModel(Long id,
