--- scaffolding-company-management-app/src/main/java/pl/com/chrzanowski/scma/repository/VehicleTireRepository.java
@@ -1,12 +1,0 @@
-package pl.com.chrzanowski.scma.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import pl.com.chrzanowski.scma.domain.VehicleTire;
-import pl.com.chrzanowski.scma.service.dto.VehicleTireDTO;
-
-@Repository
-public interface VehicleTireRepository extends JpaRepository<VehicleTire, VehicleTireDTO> {
-
-    void deleteVehicleTireById(Long id);
-}
