+++ scaffolding-company-management-app/src/main/java/pl/com/chrzanowski/scma/service/mapper/TireMapper.java
@@ -1,16 +1,17 @@
 package pl.com.chrzanowski.scma.service.mapper;
 
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
 import pl.com.chrzanowski.scma.domain.Tire;
 import pl.com.chrzanowski.scma.service.dto.TireDTO;
 
-@Mapper(componentModel = "spring", uses = {})
+@Mapper(componentModel = "spring", uses = {VehicleMapper.class})
 public interface TireMapper extends EntityMapper<TireDTO, Tire> {
 
-
+    @Mapping(source = "vehicle.id", target = "vehicleId")
     TireDTO toDto(Tire tire);
 
-
+    @Mapping(source = "vehicleId", target = "vehicle")
     Tire toEntity(TireDTO tireDTO);
