+++ scaffolding-company-management-app/src/main/resources/logback.xml
@@ -5,7 +5,6 @@
     <!-- This file contains logging to /tmp/log directory -->
     <include resource="org/springframework/boot/logging/logback/base.xml"/>
     <include resource="org/springframework/boot/logging/logback/default.xml"/>
-    <include resource="org/springframework/boot/logging/logback/console-appender.xml" />
 
     <!-- The FILE and ASYNC appenders are here as examples for a production configuration -->
